Rename events to event (#1973)
