+++ snow/event/blockable.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019-2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package events
+package event
 
 import (
 	"context"
